Update portfolio website with improved layout and new data visualisation section
This update revamps the layout of the portfolio website by adding custom styles and changing the CV section from an iframe to a canvas. A new section for Data visualisations has also been added which includes an interactive Bokeh visualisation. These new features aim to improve the user-friendliness and interactive experience for viewers.