Update image carousel and hero styling
In this commit, the image carousel in the "About Me" section of the webpage has been reordered, with one image removed. The image of "Jack Kennedy Powerlifting" has been moved to second place and "Jack and his prize in Sydney" has been positioned as the first image. Furthermore, changes have been made to the CSS, including updating the height of the hero image and adjusting the settings for carousel's width, height, and image fit.